extract out state handler
