+++ lib/core/src/lib.rs
@@ -8,6 +8,7 @@ pub mod logger;
 pub mod model;
 pub mod persist;
 pub mod sdk;
+pub(crate) mod send_swap;
 pub(crate) mod swapper;
 pub(crate) mod utils;
 pub(crate) mod wallet;
